+++ lib/src/common/color_constants.dart
@@ -5,7 +5,7 @@ abstract class ColorConstants {
 
   static const Color primary = Colors.blue;
   static Color placeholder = Colors.grey.shade300;
-  static const Color textPlaceholder = Colors.black;
+  static const Color textPlaceholder = Color(0xffBFBFBF);
   static const Color buttonDisable = Color(0xffE5E4E4);
   static const Color textButtonDisable = Color(0xffBFBFBF);
 }