+++ lib/src/common/component/calendar_picker_view_widget.dart
@@ -1,5 +1,5 @@
 import 'package:dartx/dartx.dart';
-import 'package:easy_calendar_picker/calendar_picker.dart';
+import 'package:easy_calendar_picker/easy_calendar_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';