+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.9
+
+* hotfix: remove button style color
+* replace name package
+* feat: add copyWith method for CalendarStyle
+
 ## 0.0.8
 
 * Update something